+++ lib/puppet/provider/aem_package/aem.rb
@@ -27,25 +27,16 @@ Puppet::Type.type(:aem_package).provide(:aem, parent: PuppetX::ShineSolutions::P
         max_sleep_seconds: resource[:retries_max_sleep_seconds]
       }
     }
-    check_opts = {
-      _retries: {
-        max_tries: resource[:retries_max_tries],
-        base_sleep_seconds: resource[:retries_base_sleep_seconds],
-        max_sleep_seconds: resource[:retries_max_sleep_seconds]
-      }
-    }
     package.get_versions.data.each do |version|
       package_per_version = client(resource).package(resource[:group], resource[:name], version)
       results.push(package_per_version.delete_wait_until_ready) if package_per_version.exists.data == true
     end
-    results.push(client(resource).aem.get_package_manager_servlet_status_wait_until_ready(check_opts))
-    results.push(package.create)
-    results.push(client(resource).aem.get_package_manager_servlet_status_wait_until_ready(check_opts))
-    results.push(package.update(resource[:filter]))
-    results.push(client(resource).aem.get_package_manager_servlet_status_wait_until_ready(check_opts))
-    results.push(package.build_wait_until_ready(build_opts))
-    results.push(client(resource).aem.get_package_manager_servlet_status_wait_until_ready(check_opts))
-    results.push(package.download(resource[:path]))
+
+    results.push(call_with_readiness_check(package, 'create', [], resource))
+    results.push(call_with_readiness_check(package, 'update', [resource[:filter]], resource))
+    results.push(call_with_readiness_check(package, 'build_wait_until_ready', [build_opts], resource))
+    results.push(call_with_readiness_check(package, 'download', [resource[:path]], resource))
+
     handle_multi(results)
   end
 
@@ -68,21 +59,12 @@ Puppet::Type.type(:aem_package).provide(:aem, parent: PuppetX::ShineSolutions::P
         max_sleep_seconds: resource[:retries_max_sleep_seconds]
       }
     }
-    check_opts = {
-      _retries: {
-        max_tries: resource[:retries_max_tries],
-        base_sleep_seconds: resource[:retries_base_sleep_seconds],
-        max_sleep_seconds: resource[:retries_max_sleep_seconds]
-      }
-    }
-    results.push(client(resource).aem.get_package_manager_servlet_status_wait_until_ready(check_opts))
-    results.push(package.upload_wait_until_ready(resource[:path], upload_opts))
-    results.push(client(resource).aem.get_package_manager_servlet_status_wait_until_ready(check_opts))
-    results.push(package.install_wait_until_ready(install_opts))
-    results.push(client(resource).aem.get_package_manager_servlet_status_wait_until_ready(check_opts)) if resource[:replicate] == true
-    results.push(package.replicate) if resource[:replicate] == true
-    results.push(client(resource).aem.get_package_manager_servlet_status_wait_until_ready(check_opts)) if resource[:activate] == true
-    results = results.concat(package.activate_filter(true, false)) if resource[:activate] == true
+
+    results.push(call_with_readiness_check(package, 'upload_wait_until_ready', [resource[:path], upload_opts], resource))
+    results.push(call_with_readiness_check(package, 'install_wait_until_ready', [install_opts], resource))
+    results.push(call_with_readiness_check(package, 'replicate', [], resource)) if resource[:replicate] == true
+    results = results.concat(call_with_readiness_check(package, 'activate_filter', [true, false], resource)) if resource[:activate] == true
+
     handle_multi(results)
   end
 
@@ -90,8 +72,10 @@ Puppet::Type.type(:aem_package).provide(:aem, parent: PuppetX::ShineSolutions::P
   def destroy
     package = client(resource).package(resource[:group], resource[:name], resource[:version])
     results = []
+
     results.push(package.uninstall) if package.is_installed.data
     results.push(package.delete)
+
     handle_multi(results)
   end